refactor(automation): consistent file naming (handlers_*/engine_*) and rewrite the docs
The automation module mixed naming schemes that read as accidental
synonyms and made "how many workflows are there and where" hard to scan.
Regroup the module's files onto a clear per-role scheme, no logic change:

- handlers_event.go / handlers_clock.go: the shipped handlers (the four
  bus-triggered + three clock-triggered; the old workflows_starter.go +
  workflows_event_handlers.go merged, workflows_clock_handlers.go renamed)
- engine.go / engine_run.go / engine_blocked.go: the runtime (was
  workflows.go / workflows_run.go / workflows_blocked.go)
- catalog_*.go / automations_*.go: the registry and the store (unchanged)

The word "workflow" now survives only where it is the spec seam or a
shipped surface — the frozen shared/ports/workflow port, the WorkflowEngine
type, the workflow_run table, the cg:workflows group, compose/workflows.go
(the composition wiring, deliberately left as-is). ~70 stale comment/doc
references swept across code and docs, disambiguating module-engine
references from the compose sibling.

gen-workflow now scaffolds handlers_<name>.go instead of
workflows_<name>.go, so new handlers stay consistent (tool + test updated).

Docs:
- rewrite docs/explanation/automation.md in the write-backbone.md style:
  onboarding-first, a glance diagram then numbered deep dives, with the
  concrete 7 triggers / 7 actions as tables rather than an abstract bullet,
  and an explicit "automation (product) vs workflow (engine)" glossary rule
  anchored to the spec/contract.
- rename docs/how-to/add-an-automation-trigger-or-action.md to
  create-a-workflow.md and retitle it "Create an automation workflow"
  (the primary case it already documented); update index links.
- doc.go gains a pointer to compose/workflows.go as the registration
  source of truth.

No runtime behavior changes; reviewed clean by craft static plus the
craft-reviewer and security-redteam agents.
