+++ backend/internal/shared/ports/workflow/workflow.go
@@ -52,7 +52,7 @@ type Spec struct {
 
 // Trigger binds to the event bus or a schedule: EventType for bus events,
 // Schedule when EventType is empty — a clock:<name> marker, never a cron
-// expression (modules/automation/workflows_clock_handlers.go's
+// expression (modules/automation/handlers_clock.go's
 // noActivityScheduleMarker doc): the real cadence is the River periodic
 // job's own interval, and a Schedule-bearing handler also needs its own
 // candidate source wired at the time-scan (modules/automation/
